test-requirements.txt
flake8
mock
pep8
pytest-cov
pytest-localserver
